refactor: split server.ts God module into focused sub-modules (#838)
server.ts (3589 → 200 lines):
- Extract LocalBackend class + types/helpers → backend.ts (991 lines)
- Extract 29 tool definitions → tools.ts (1924 lines)
- Extract resource handlers → resources.ts (358 lines)
- Extract prompt handlers → prompts.ts (166 lines)

Uses factory pattern (createTools, createResourceHandlers) for modules
that close over the backend singleton. No behavioral changes — pure
structural refactor. All 129 MCP tests pass.
